refactor: address code quality issues while preserving multi-value review
- Extract large mock data (540+ lines) to /src/lib/mock-case-details.ts
- Create constants file for magic strings (AI model names, etc.)
- Break down 2000+ line component into smaller, focused components:
  - CaseHeader: Navigation and title display
  - DocumentsList: Document management section
  - DataField: Individual field display with multi-value support
  - DataFieldSection: All data sections organized by category
  - EditDialog: Field editing functionality
- Add proper TypeScript types in dedicated types file
- Preserve multi-value review functionality for AI model extractions
- Reduce main component to ~200 lines while maintaining all features
- Follow single responsibility principle throughout
