+++ src/app/admin/cases/[id]/components/index.ts
@@ -1,0 +1,5 @@
+export { CaseHeader } from './CaseHeader';
+export { DataField } from './DataField';
+export { DataFieldSection } from './DataFieldSection';
+export { DocumentsList } from './DocumentsList';
+export { EditDialog } from './EditDialog';
